extract prop types, helper function
